Refactor auth context and add Supabase client
